requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
idna==3.6
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.2.1
